anms-core/pyproject.toml
[build-system]
requires = ["setuptools>=62.3"]
build-backend = "setuptools.build_meta"

[project]
name = "anms-core"
version = "0.0.0"
authors = [
]
description = "anms-core package"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
  "anms-ace",
  "anms-camp",
  "anyio ~=3.6.1",
  "arrow ~=1.2.3",
  "asyncio ~=3.4.3",
  "asyncpg ~=0.27.0",
  "authlib ~=0.15.5",
  "docker ~=6.0.1",
  "emails ~=0.6",
  "email-validator ~=1.3",
  "fastapi ~=0.86.0",
  "fastapi-pagination ~=0.9.1",
  "grpcio ~=1.50.0",
  "grpcio-tools ~=1.50.0",
  "gunicorn ~=20.1.0",
  "httpx ~=0.24.0",
  "itsdangerous ~=2.1.2",
  "jinja2 ~=3.1.2",
  "munch",
  "opensearch-py ~=2.0.0",
  "paho-mqtt ~=1.6.1",
  "psycopg2-binary ~=2.9.5",
  "pydash ~=5.1.1",
  "pydantic",
  "pyotp ~=2.7.0",
  "sqlalchemy ~=1.4.42",
  "starlette",
  "starlette-early-data ~=1.1.0",
  "tenacity ~=8.0.1",
  "trio ~=0.20.0",
  "uvicorn[standard] ~=0.19.0",
  "python-multipart ~=0.0.5"
]

[project.optional-dependencies]
test = [
  "pytest ~=7.1.2",
  "pytest-anyio",
  "pytest-mock ~=3.7.0",
  "pytest-cov ~=3.0.0",
  "mock ~=4.0.3",
  "mock-alchemy ==0.2.5",
  "coverage ~=6.5.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["anms*"]

[tool.setuptools.package-data]
anms = [
  '**/*.html',
  '**/*.css',
  '**/*.js',
]
